Change Review Response

+++ src/main/java/com/backend/category/CategoryService.java
@@ -2,9 +2,12 @@ package com.backend.category;
 
 import com.backend.Error;
 import com.backend.category.dto.CreateCategoryDTO;
+import com.backend.movie.MovieEntity;
 import com.backend.security.JwtUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Service;
 
@@ -63,6 +66,12 @@ public class CategoryService {
         if(existedCategory == null){
             throw Error.NotFoundError("Category");
         }
+
+        Query query = new Query(Criteria.where("categories").is(existedCategory.getId()));
+        List<MovieEntity> moviesOfCate = mongoTemplate.find(query, MovieEntity.class);
+        if(moviesOfCate.size() > 0)
+            throw Error.UsedError("Category");
+
         categoryRespository.deleteById(id);
 
         return existedCategory;
